requirements.txt
jupyter~=1.0.0
style~=1.1.0
pip~=23.1.1
attrs~=22.1.0
wheel~=0.42.0
cryptography~=40.0.2
Pillow~=9.2.0
tornado~=6.3.3
Jinja2~=3.1.2
cffi~=1.15.1
binance~=0.3
requests~=2.28.2
websocket-client~=1.6.1
certifi~=2020.6.20
urllib3~=1.25.10
psycopg2~=2.9.6
idna~=2.10
chardet~=3.0.4
websockets~=11.0.2
setuptools~=67.7.1
packaging~=23.1
pyparsing~=2.4.7
zipp~=3.9.0